xc refactor : 1. add xc_funcs.h for compiling without libxc 2. removed some no longer used subroutines
