+++ source/module_xc/CMakeLists.txt
@@ -1,6 +1,5 @@
 list(APPEND objects
     H_XC_pw.cpp
-    potential_libxc.cpp
     potential_libxc_meta.cpp
     xc_functional.cpp
     xc_gga_pw.cpp
