--- source/module_xc/potential_libxc.h
@@ -1,27 +1,0 @@
-//==========================================================
-// AUTHOR : Peize Lin
-// DATE :   2017-09-14
-// UPDATE : 2021-02-28
-//==========================================================
-
-#ifdef USE_LIBXC
-
-#ifndef POTENTIAL_LIBXC_H
-#define POTENTIAL_LIBXC_H
-
-#include "../module_base/matrix.h"
-#include "../module_base/vector3.h"
-#include "../module_base/global_variable.h"
-#include "../module_base/global_function.h"
-#include <xc.h>
-#include <vector>
-#include <tuple>
-
-class Potential_Libxc
-{
-
-};
-
-#endif
-
-#endif	// ifdef USE_LIBXC
