+++ source/module_xc/xc_funcs.h
@@ -1,0 +1,21 @@
+#define  XC_LDA_X                         1  /*Exchange                                                              */
+#define  XC_LDA_C_PZ                      9  /*Perdew & Zunger                                                       */
+#define  XC_LDA_C_PW                     12  /*Perdew & Wang                                                         */
+#define  XC_HYB_GGA_XC_PBEH             406  /*aka PBE0 or PBE1PBE                                                   */
+#define  XC_HYB_GGA_XC_HSE06            428  /* the 2006 version of the screened hybrid HSE                          */
+#define  XC_GGA_X_HCTH_A                 34  /*HCTH-A                                                                */
+#define  XC_GGA_C_HCTH_A                 97  /*HCTH-A                                                                */
+#define  XC_GGA_X_PBE                   101  /*Perdew, Burke & Ernzerhof exchange                                    */
+#define  XC_GGA_X_B88                   106  /*Becke 88                                                              */
+#define  XC_GGA_X_PW91                  109  /*Perdew & Wang 91                                                      */
+#define  XC_GGA_X_OPTX                  110  /*Handy & Cohen OPTX 01                                                 */
+#define  XC_GGA_X_PBE_SOL               116  /* Perdew, Burke & Ernzerhof exchange (solids)                          */
+#define  XC_GGA_X_RPBE                  117  /*Hammer, Hansen & Norskov (PBE-like)                                   */
+#define  XC_GGA_X_WC                    118  /*Wu & Cohen                                                            */
+#define  XC_GGA_C_PBE                   130  /*Perdew, Burke & Ernzerhof correlation                                 */
+#define  XC_GGA_C_LYP                   131  /*Lee, Yang & Parr                                                      */
+#define  XC_GGA_C_P86                   132  /*Perdew 86                                                             */
+#define  XC_GGA_C_PBE_SOL               133  /* Perdew, Burke & Ernzerhof correlation SOL                            */
+#define  XC_GGA_C_PW91                  134  /*Perdew & Wang 91                                                      */
+#define  XC_MGGA_X_SCAN                 263  /*SCAN exchange of Sun, Ruzsinszky, and Perdew                          */
+#define  XC_MGGA_C_SCAN                 267  /*SCAN correlation                                                      */
