+++ source/module_xc/xc_functional.h
@@ -7,6 +7,8 @@
 
 #ifdef USE_LIBXC
 #include <xc.h>
+#else
+#include "xc_funcs.h"
 #endif	// ifdef USE_LIBXC
 
 #include "../module_base/global_function.h"
@@ -37,43 +39,11 @@ class XC_Functional
 		const double*const*const rho_in, 
 		const double*const rho_core); // core charge density
 
-	//------------------------------------------------
-	// evaluate the exchange-correlation (XC) energy
-	// by using the input charge density rho_in and rho_core_in
-	//------------------------------------------------
-	// [etxc, vtxc, v] = v_xc(...)
-	static std::tuple<double,double,ModuleBase::matrix> v_xc_libxc(
-		const int &nrxx, // number of real-space grid
-		const int &ncxyz, // total number of charge grid
-		const double &omega, // volume of cell
-		const double * const * const rho_in,
-		const double * const rho_core_in);
-	
 	static std::tuple<double,double,ModuleBase::matrix,ModuleBase::matrix> v_xc_meta(
 		const double * const * const rho_in,
 		const double * const rho_core_in,
 		const double * const * const kin_r_in);
-		
-#ifdef USE_LIBXC
-	struct Mgga_spin_in
-	{
-		double rhoup, rhodw;//electron densities
-		ModuleBase::Vector3<double> grhoup, grhodw;//gradient of electron densities
-		double tauup, taudw;//kinetic energy densities
-	};
 
-	struct Mgga_spin_out
-	{
-		double ex, ec;//xc energy densities
-		double v1xup, v1xdw;//vx: lda part
-		double v2xup, v2xdw;//vx: gga part
-		double v3xup, v3xdw;//vx: mgga part
-		double v1cup, v1cdw;//vc: lda part
-		ModuleBase::Vector3<double> v2cup, v2cdw;
-		std::vector<double> v2c;//vc: gga part, two different formats	
-		double v3cup, v3cdw;//vc: mgga part
-	};	
-	
 	// GGA
 	static void gcxc(const double &rho, const double &grho,
 			double &sxc, double &v1xc, double &v2xc);
@@ -99,11 +69,29 @@ class XC_Functional
 	static void noncolin_rho(double *rhoout1,double *rhoout2,double *seg);
 
 	// mGGA
+#ifdef USE_LIBXC
+	struct Mgga_spin_in
+	{
+		double rhoup, rhodw;//electron densities
+		ModuleBase::Vector3<double> grhoup, grhodw;//gradient of electron densities
+		double tauup, taudw;//kinetic energy densities
+	};
+
+	struct Mgga_spin_out
+	{
+		double ex, ec;//xc energy densities
+		double v1xup, v1xdw;//vx: lda part
+		double v2xup, v2xdw;//vx: gga part
+		double v3xup, v3xdw;//vx: mgga part
+		double v1cup, v1cdw;//vc: lda part
+		ModuleBase::Vector3<double> v2cup, v2cdw;
+		std::vector<double> v2c;//vc: gga part, two different formats	
+		double v3cup, v3cdw;//vc: mgga part
+	};	
 	static void tau_xc(const double &rho, const double &grho, const double &atau, double &sx, double &sc,
           double &v1x, double &v2x, double &v3x, double &v1c, double &v2c, double &v3c);
 	static void tau_xc_spin(const Mgga_spin_in &mgga_spin_in, Mgga_spin_out &mgga_spin_out);
-
-#endif
+#endif 
 
 	static int get_func_type();
 
@@ -114,8 +102,9 @@ class XC_Functional
 	static bool use_libxc;
 
 	static void set_xc_type(const std::string xc_func_in);
+#ifdef USE_LIBXC
 	static std::vector<xc_func_type> init_func(const int xc_polarized);
-
+#endif
 	// LDA
 	static void xc(const double &rho, double &exc, double &vxc);
 	static void xc_libxc(const double &rho, double &exc, double &vxc);
