+++ source/src_lcao/FORCE_STRESS.cpp
@@ -1,6 +1,5 @@
 #include "FORCE_STRESS.h"
 #include "../src_pw/global.h"
-#include "../module_xc/potential_libxc.h"
 #include "./dftu.h"  //Quxin add for DFT+U on 20201029
 // new
 #include "../src_pw/vdwd2.h"
