+++ source/src_pw/charge.cpp
@@ -47,7 +47,7 @@ Charge::~Charge()
 			delete[] rhog[i];
 			delete[] rho_save[i];
 			delete[] rhog_save[i];
-			if(GlobalV::DFT_META)
+			if(XC_Functional::get_func_type() == 3)
 			{
 				delete[] kin_r[i];
 				delete[] kin_r_save[i];
@@ -59,7 +59,7 @@ Charge::~Charge()
 		delete[] rhog_save;
     	delete[] rho_core;
 		delete[] rhog_core;
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
 			delete[] kin_r;
 			delete[] kin_r_save;
@@ -90,7 +90,7 @@ void Charge::allocate(const int &nspin_in, const int &nrxx_in, const int &ngmc_i
 	rhog = new std::complex<double>*[nspin];
 	rho_save = new double*[nspin];
 	rhog_save = new std::complex<double>*[nspin];
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
 		kin_r = new double*[nspin];
 		kin_r_save = new double*[nspin];
@@ -106,7 +106,7 @@ void Charge::allocate(const int &nspin_in, const int &nrxx_in, const int &ngmc_i
 		ModuleBase::GlobalFunc::ZEROS(rhog[is], ngmc);
 		ModuleBase::GlobalFunc::ZEROS(rho_save[is], nrxx);
 		ModuleBase::GlobalFunc::ZEROS(rhog_save[is], ngmc);
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
 			kin_r[is] = new double[nrxx];
 			ModuleBase::GlobalFunc::ZEROS(kin_r[is], nrxx);
@@ -511,7 +511,7 @@ void Charge::atomic_rho(const int spin_number_need, double** rho_in)const		// Pe
 			rho_in[is][ir] = rho_in[is][ir] / ne_tot * nelec;
 
 	//wenfei 2021-7-29 : initial tau = 3/5 rho^2/3, Thomas-Fermi
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
 		const double pi = 3.141592653589790;
 		double fact = (3.0/5.0)*pow(3.0*pi*pi,2.0/3.0);
@@ -725,7 +725,7 @@ void Charge::sum_band(void)
 	for(int is=0; is<GlobalV::NSPIN; is++)
 	{
 		ModuleBase::GlobalFunc::ZEROS(rho[is], GlobalC::pw.nrxx);
-		if (GlobalV::DFT_META)
+		if (XC_Functional::get_func_type() == 3)
 		{
 			ModuleBase::GlobalFunc::ZEROS(kin_r[is], GlobalC::pw.nrxx);	
 		}
@@ -840,7 +840,7 @@ void Charge::sum_band_k(void)
 			}
 
 			//kinetic energy density
-			if (GlobalV::DFT_META)
+			if (XC_Functional::get_func_type() == 3)
 			{
 				for (int j=0; j<3; j++)
 				{
@@ -973,7 +973,7 @@ void Charge::rho_mpi(void)
 	double *tau_tot;
 	double *tau_tot_aux;
 
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
     	tau_tmp = new double[GlobalC::pw.nrxx];
 	    tau_tot = new double[GlobalC::pw.ncxyz];
@@ -984,19 +984,19 @@ void Charge::rho_mpi(void)
     for (int is=0; is< GlobalV::NSPIN; is++)
     {
         ModuleBase::GlobalFunc::ZEROS(rho_tot, GlobalC::pw.ncxyz);
-		if(GlobalV::DFT_META) ModuleBase::GlobalFunc::ZEROS(tau_tot, GlobalC::pw.ncxyz);
+		if(XC_Functional::get_func_type() == 3) ModuleBase::GlobalFunc::ZEROS(tau_tot, GlobalC::pw.ncxyz);
 
 		for (ir=0;ir<GlobalC::pw.nrxx;ir++)
 		{
 			rho_tmp[ir] = this->rho[is][ir] / static_cast<double>(GlobalV::NPROC_IN_POOL);
-			if(GlobalV::DFT_META)
+			if(XC_Functional::get_func_type() == 3)
 			{
 				tau_tmp[ir] = this->kin_r[is][ir] / static_cast<double>(GlobalV::NPROC_IN_POOL);
 			}
 		}
 
         MPI_Allgatherv(rho_tmp, GlobalC::pw.nrxx, MPI_DOUBLE, rho_tot, rec, dis, MPI_DOUBLE, POOL_WORLD);
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
         	MPI_Allgatherv(tau_tmp, GlobalC::pw.nrxx, MPI_DOUBLE, tau_tot, rec, dis, MPI_DOUBLE, POOL_WORLD);
 		}
@@ -1005,7 +1005,7 @@ void Charge::rho_mpi(void)
         // this is the most complicated part !!
         //=================================================================
         ModuleBase::GlobalFunc::ZEROS(rho_tot_aux, GlobalC::pw.ncxyz);
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
         	ModuleBase::GlobalFunc::ZEROS(tau_tot_aux, GlobalC::pw.ncxyz);
 		}
@@ -1048,7 +1048,7 @@ void Charge::rho_mpi(void)
 					// -------------------------------------------------
                     rho_tot_aux[GlobalC::pw.ncz*ir    + start_z[ip]      + iz]
                       = rho_tot[num_z[ip]*ir + start_z[ip]*ncxy + iz];
-					if(GlobalV::DFT_META)
+					if(XC_Functional::get_func_type() == 3)
 					{
                     	tau_tot_aux[GlobalC::pw.ncz*ir    + start_z[ip]      + iz]
                       	  = tau_tot[num_z[ip]*ir + start_z[ip]*ncxy + iz];
@@ -1062,14 +1062,14 @@ void Charge::rho_mpi(void)
 		if(GlobalV::CALCULATION=="scf-sto" || GlobalV::CALCULATION=="relax-sto" || GlobalV::CALCULATION=="md-sto") //qinarui add it temporarily.
 		{
 			MPI_Allreduce(rho_tot_aux,rho_tot,GlobalC::pw.ncxyz,MPI_DOUBLE,MPI_SUM,POOL_WORLD);
-			if(GlobalV::DFT_META)
+			if(XC_Functional::get_func_type() == 3)
 			{
 				MPI_Allreduce(tau_tot_aux,tau_tot,GlobalC::pw.ncxyz,MPI_DOUBLE,MPI_SUM,POOL_WORLD);
 			}
 		}
 		else
         MPI_Allreduce(rho_tot_aux,rho_tot,GlobalC::pw.ncxyz,MPI_DOUBLE,MPI_SUM,MPI_COMM_WORLD);
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
    	    	MPI_Allreduce(tau_tot_aux,tau_tot,GlobalC::pw.ncxyz,MPI_DOUBLE,MPI_SUM,MPI_COMM_WORLD);
 		}
@@ -1082,7 +1082,7 @@ void Charge::rho_mpi(void)
             for (iz=0;iz<num_z[GlobalV::RANK_IN_POOL];iz++)
             {
                 this->rho[is][num_z[GlobalV::RANK_IN_POOL]*ir+iz] = rho_tot[GlobalC::pw.ncz*ir + start_z[GlobalV::RANK_IN_POOL] + iz ];
-				if(GlobalV::DFT_META)
+				if(XC_Functional::get_func_type() == 3)
 				{
                 	this->kin_r[is][num_z[GlobalV::RANK_IN_POOL]*ir+iz] = tau_tot[GlobalC::pw.ncz*ir + start_z[GlobalV::RANK_IN_POOL] + iz ];
 				}	
@@ -1093,7 +1093,7 @@ void Charge::rho_mpi(void)
     delete[] rho_tot;
     delete[] rho_tmp;
     
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
 		delete[] tau_tot_aux;
     	delete[] tau_tot;
@@ -1115,7 +1115,7 @@ void Charge::save_rho_before_sum_band(void)
 	for(int is=0; is<GlobalV::NSPIN; is++)
 	{
     	ModuleBase::GlobalFunc::DCOPY( rho[is], rho_save[is], GlobalC::pw.nrxx);
-    	if(GlobalV::DFT_META) ModuleBase::GlobalFunc::DCOPY( kin_r[is], kin_r_save[is], GlobalC::pw.nrxx);
+    	if(XC_Functional::get_func_type() == 3) ModuleBase::GlobalFunc::DCOPY( kin_r[is], kin_r_save[is], GlobalC::pw.nrxx);
     }
     return;
 }
