+++ source/src_pw/energy.cpp
@@ -388,7 +388,7 @@ double energy::delta_e(void)
     for (int ir=0; ir<GlobalC::pw.nrxx; ir++)
     {
     	deband_aux -= GlobalC::CHR.rho[0][ir] * GlobalC::pot.vr(0, ir);
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
 			deband_aux -= GlobalC::CHR.kin_r[0][ir] * GlobalC::pot.vofk(0,ir);
 		}
@@ -399,7 +399,7 @@ double energy::delta_e(void)
     	for (int ir=0; ir<GlobalC::pw.nrxx; ir++)
     	{
     		deband_aux -= GlobalC::CHR.rho[1][ir] * GlobalC::pot.vr(1, ir);
-			if(GlobalV::DFT_META)
+			if(XC_Functional::get_func_type() == 3)
 			{
 				deband_aux -= GlobalC::CHR.kin_r[1][ir] * GlobalC::pot.vofk(1,ir);
 			}
@@ -444,7 +444,7 @@ void energy::delta_escf(void)
     for (int ir=0; ir<GlobalC::pw.nrxx; ir++)
     {
 		this->descf -= ( GlobalC::CHR.rho[0][ir] - GlobalC::CHR.rho_save[0][ir] ) * GlobalC::pot.vr(0, ir);
-		if(GlobalV::DFT_META)
+		if(XC_Functional::get_func_type() == 3)
 		{
          	this->descf -= ( GlobalC::CHR.kin_r[0][ir] - GlobalC::CHR.kin_r_save[0][ir] ) * GlobalC::pot.vofk(0, ir);
 		}
@@ -455,7 +455,7 @@ void energy::delta_escf(void)
        	for (int ir=0; ir<GlobalC::pw.nrxx; ir++)
        	{
            	this->descf -= ( GlobalC::CHR.rho[1][ir] - GlobalC::CHR.rho_save[1][ir] ) * GlobalC::pot.vr(1, ir);
-			if(GlobalV::DFT_META)
+			if(XC_Functional::get_func_type() == 3)
 			{
            		this->descf -= ( GlobalC::CHR.kin_r[1][ir] - GlobalC::CHR.kin_r_save[1][ir] ) * GlobalC::pot.vofk(1, ir);
 			}
