+++ source/src_pw/hamilt_pw_hip.cpp
@@ -1578,7 +1578,7 @@ void Hamilt_PW::h_psi(const std::complex<double> *psi_in, std::complex<double> *
 	// (4) the metaGGA part
 	//------------------------------------
 	// timer::tick("Hamilt_PW","meta");
-	if (GlobalV::DFT_META)
+	if (XC_Functional::get_func_type() == 3)
 	{
 		tmhpsi = hpsi;
 		tmpsi_in = psi_in;
