+++ source/src_pw/potential.cpp
@@ -6,7 +6,6 @@
 #include "../module_xc/xc_functional.h"
 #include "efield.h"
 #include "math.h"
-#include "../module_xc/potential_libxc.h"
 // new
 #include "H_Hartree_pw.h"
 #ifdef __LCAO
@@ -44,7 +43,7 @@ void Potential::allocate(const int nrxx)
     ModuleBase::Memory::record("Potential","vr",GlobalV::NSPIN*nrxx,"double");
     ModuleBase::Memory::record("Potential","vr_eff",GlobalV::NSPIN*nrxx,"double");
 	
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
 		this->vofk.create(GlobalV::NSPIN,nrxx);
     	ModuleBase::Memory::record("Potential","vofk",GlobalV::NSPIN*nrxx,"double");
@@ -82,7 +81,7 @@ void Potential::init_pot(
     // the vltot should and must be zero here.
     ModuleBase::GlobalFunc::ZEROS(this->vltot, GlobalC::pw.nrxx);
 
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
 		this->vofk.zero_out();
 	}
@@ -325,14 +324,15 @@ ModuleBase::matrix Potential::v_of_rho(
 //  calculate the exchange-correlation potential
 //----------------------------------------------------------
 	
-	#ifdef USE_LIBXC
-	if(GlobalV::DFT_META)
+	if(XC_Functional::get_func_type() == 3)
 	{
+#ifdef USE_LIBXC
     	const std::tuple<double,double,ModuleBase::matrix,ModuleBase::matrix> etxc_vtxc_v = XC_Functional::v_xc_meta(rho_in, GlobalC::CHR.rho_core, GlobalC::CHR.kin_r);
 		GlobalC::en.etxc = std::get<0>(etxc_vtxc_v);
 		GlobalC::en.vtxc = std::get<1>(etxc_vtxc_v);
 		v            += std::get<2>(etxc_vtxc_v);
-		vofk		  = std::get<3>(etxc_vtxc_v);	
+		vofk		  = std::get<3>(etxc_vtxc_v);
+#endif
 	}
 	else
 	{	
@@ -341,13 +341,6 @@ ModuleBase::matrix Potential::v_of_rho(
 		GlobalC::en.vtxc = std::get<1>(etxc_vtxc_v);
 		v            += std::get<2>(etxc_vtxc_v);
 	}
-	#else
-	const std::tuple<double,double,ModuleBase::matrix> etxc_vtxc_v = XC_Functional::v_xc(GlobalC::pw.nrxx, GlobalC::pw.ncxyz, GlobalC::ucell.omega, rho_in, GlobalC::CHR.rho_core);
-	
-	GlobalC::en.etxc = std::get<0>(etxc_vtxc_v);
-	GlobalC::en.vtxc = std::get<1>(etxc_vtxc_v);
-	v            += std::get<2>(etxc_vtxc_v);
-	#endif
 
 //----------------------------------------------------------
 //  calculate the Hartree potential
